+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.1604.1 (unreleased)
+# v0.1605.0
 
 ## Additions